inventory view cache: moving tests to its own file
